[#92] fixed acceleration structure creation.
I revisited the acceleration structure creation and tried different things and now it finally works (not perfect, but it works). The main issue was (i think) the data being given as uint8_t. I changed it to float for vertices and uint32_t for indices and now it works. Additionally i rewrote the way we submit commands to the commandbuffer.

+++ modules/rtx/include/vkcv/rtx/RTX.hpp
@@ -27,8 +27,8 @@ namespace vkcv::rtx {
          * @param indices The scene index data of type uint8_t.
          * @param descriptorSetHandles The descriptorSetHandles for RTX
          */
-        RTXModule(Core* core, ASManager* asManager, std::vector<uint8_t>& vertices,
-            std::vector<uint8_t>& indices, std::vector<vkcv::DescriptorSetHandle>& descriptorSetHandles);
+        RTXModule(Core* core, ASManager* asManager, std::vector<float>& vertices,
+            std::vector<uint32_t>& indices, std::vector<vkcv::DescriptorSetHandle>& descriptorSetHandles);
 
         /**
          * @brief Default #RTXModule destructor.
